requirements.txt
langchain
langchain-openai
manim
Pillow
python-dotenv
numpy
opencv-python
streamlit